backend/requirements.txt
Flask
flask-jwt-extended
python-dotenv
pymysql
werkzeug
